+++ client/src/timeline/Main.js
@@ -1,14 +1,19 @@
 import React from 'react';
-import Posts from './Post';
+import Grid from '@material-ui/core/Grid';
+import Posts from './Posts';
 import PostForm from './PostForm';
 
 const Timeline = (props) => {
-  console.log(props.username);
   return (
-    <div>
-      <PostForm />
-      <Posts username={props.username} />
-    </div>
+    <Grid container>
+      <Grid item xs={12} sm={3}>
+        menu
+      </Grid>
+      <Grid item xs={12} sm={9}>
+        <PostForm />
+        <Posts username={props.username}/>
+      </Grid>
+    </Grid>
   );
 }
