+++ client/src/timeline/Post.js
@@ -1,18 +1,27 @@
 import React from 'react';
+import { withStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
+import CardHeader from '@material-ui/core/CardHeader';
+import Avatar from '@material-ui/core/Avatar';
 import CardActions from '@material-ui/core/CardActions';
 import CardContent from '@material-ui/core/CardContent';
 import Typography from '@material-ui/core/Typography';
 import Icon from '@material-ui/core/Icon';
-import { Redirect } from 'react-router-dom';
+import IconButton from '@material-ui/core/IconButton';
 import axios from 'axios';
 
+const styles = {
+  root: {
+    minWidth: "95%",
+  },
+};
 
 class Like extends React.Component {
   constructor(props){
     super(props);
     this.state = {
       like: props.like,
+      num: props.num,
     }
     this.handleClick = this.handleClick.bind(this);
   }
@@ -23,93 +32,45 @@ class Like extends React.Component {
       .then((res) => {
         this.setState({
           like: !this.state.like,
+          num: this.state.num + (this.state.like ? -1 : 1),
         });
       });
   }
 
   render() {
     return (
-      <span onClick={this.handleClick}>
-          { this.state.like ? <Icon>{'favorite'}</Icon> : <Icon>{'favorite_border'}</Icon> }
+      <span>
+        <IconButton onClick={this.handleClick}>
+            { this.state.like ? <Icon>{'favorite'}</Icon> : <Icon>{'favorite_border'}</Icon> }
+        </IconButton>
+        { this.state.num }
       </span>
     );
   }
 }
 
 class Post extends React.Component {
-
   render() {
     return (
-      <Card key={this.props.postId}>
+      <Card key={this.props.postId} className={this.props.classes.root}>
+        <CardHeader
+          avatar={
+            <Avatar alt={this.props.username}>{this.props.username.charAt(0)}</Avatar>
+          }
+          title={this.props.username}
+          subheader={this.props.date.toString()}
+        />
         <CardContent>
-          <Typography color='textSecondary' gutterBottom>
-            { this.props.username }
-          </Typography>
           <Typography variant='body2' component='p'>
             { this.props.text }
           </Typography>
         </CardContent>
         <CardActions>
-          <Like like={this.props.isLike} postId={this.props.postId} />
+          <Like like={this.props.likes.indexOf(this.props.me) >= 0} num={this.props.likes.length} postId={this.props.postId} />
         </CardActions>
       </Card>
     );
   }
 }
 
-export default class Posts extends React.Component {
-  constructor(props)
-  {
-    console.log(props.username);
-    super(props);
-    this.state = {
-      posts: [],
-      redirect: false,
-    };
-    this.tick();
-    setInterval(() => this.tick(), 10 * 1000);
-
-    this.componentDidUpdate = this.componentDidUpdate.bind(this);
-  }
-  
-
-  tick() {
-    if(!this.props.username) return;
-    axios
-      .get('/api/post')
-      .then((res) => {
-        if(res.status !== 200) return;
-        this.setState({
-          posts: res.data,
-        });
-      })
-      .catch((err) => {
-        console.log(err);
-        if(err.response.status === 401) {
-          this.setState({
-            redirect: true,
-          });
-        }
-      });
-  }
-
-  componentDidUpdate(prevProps) {
-    if(prevProps.username !== this.props.username) {
-      this.tick();
-    }
-  }
-
-  render(){
-    return (
-      <ul>
-        { 
-          this.state.redirect
-          ? (<Redirect to='/signin' />)
-          : this.state.posts.map(post => {
-            return <Post key={post.id} username={post.username} postId={post.id} text={post.text} isLike={(post.likes.indexOf(this.props.username) >= 0)} me={this.props.username} />
-          })
-        }
-      </ul>
-    );
-  }
-}
+export default withStyles(styles)(Post);
